@@ -11,10 +11,10 @@
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "ahooks": "^3.7.0",
-    "antd-mobile": "5.9.0",
+    "antd-mobile": "^5.21.0",
     "axios": "^0.27.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "7",
     "react-router-dom": "5.3.0",
     "react-scripts": "5.0.1",
