@@ -1,13 +1,13 @@
-import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import App from './App';
-import './index.scss';
-import store from './store';
+import ReactDOM from 'react-dom/client'
+import { Provider } from 'react-redux'
+import App from './App'
+import './index.scss'
+import store from './store'
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
+
+root.render(
   <Provider store={store}>
     <App />
-  </Provider>,
-   document.getElementById('root')
-);
-
+  </Provider>
+)
