@@ -1,9 +1,11 @@
 import { combineReducers } from 'redux'
 
 import login from './login'
+import profile from './profile'
 
 const rootReducers = combineReducers({
-  login
+  login,
+  profile
 })
 
 
